Update blueprint support to latest trunk.
With an extra hack to suppress "source file not found errors", because the
tests still need updating for this to change.

Flask-Assets used to ignore missing files (different from the webassets core),
but this was always an oversight, and is now extra effort to implement, so
we will get rid of the old behaviour in a future commit.