Settle use of webassets load_path/directory.
Mostly, when the webassets load_path system should be used, when webassets
Environment.directory and Environment.url should be used, and when to use
the Flask static folder system is now put in writing.

The actual implementation and behaviour has changed slightly and subtly to
follow what the design sets out to do.

Various tests needed to be updated along the way.

Closes #30.